Architecture completed

+++ src/main/java/org/app/AirConditioningApplication/Service/BudgetService.java
@@ -3,10 +3,12 @@ package org.app.AirConditioningApplication.Service;
 import org.app.AirConditioningApplication.Model.Budget;
 import org.app.AirConditioningApplication.Repository.BudgetRepo;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+@Service
 public class BudgetService {
     private final BudgetRepo budgetRepo;
 
@@ -40,7 +42,9 @@ public class BudgetService {
     public ResponseEntity<Object> getById(Long Id) {
         try {
             Optional<Budget> budget = budgetRepo.findById(Id);
-            return ResponseEntity.ok().body(budget);
+            if (budget.isPresent())
+                return ResponseEntity.ok().body(budget);
+            else return ResponseEntity.ok().body("Invalid Id");
         } catch (Exception e) {
             return ResponseEntity.ok().body(e.getMessage());
         }
@@ -50,8 +54,11 @@ public class BudgetService {
     public ResponseEntity<Object> delete(Long Id) {
         try {
             Optional<Budget> budget = budgetRepo.findById(Id);
-            budgetRepo.delete(budget.get());
-            return ResponseEntity.ok().body("Deleted");
+            if(budget.isPresent()){
+                budgetRepo.delete(budget.get());
+                return ResponseEntity.ok().body("Deleted");
+            }else return ResponseEntity.ok().body("Invalid ID");
+
         } catch (Exception e) {
             return ResponseEntity.ok().body(e.getMessage());
         }
